+++ src/main/java/info/u_team/u_team_core/item/tool/UPickaxeItem.java
@@ -15,7 +15,7 @@ public class UPickaxeItem extends PickaxeItem implements IURegistryType {
 	}
 	
 	public UPickaxeItem(String name, ItemGroup group, Properties properties, IToolMaterial material) {
-		super(material, material.getAdditionalDamage(Tools.PICKAXE), material.getAttackSpeed(Tools.PICKAXE), applyToolType(group == null ? properties : properties.group(group), material));
+		super(material, (int) material.getAdditionalDamage(Tools.PICKAXE), material.getAttackSpeed(Tools.PICKAXE), applyToolType(group == null ? properties : properties.group(group), material));
 		this.name = name;
 	}
